* Fixed архітектуру
