+++ src/main/java/animals/Cat.java
@@ -1,6 +1,6 @@
 package animals;
 
-public class Cat extends Animal{
+ public class Cat extends Animal{
 
     public Cat()
     {
