+++ src/test/java/animals/cat/Test1.java
@@ -1,0 +1,16 @@
+package animals.cat;
+
+import animals.Cat;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+public class Test1 {
+
+    @Test
+    private void testCatVoice()
+    {
+        Cat shredingerCat = new Cat();
+        Assert.assertEquals(shredingerCat.getVoice(), "Мяв!");
+    }
+
+}
