+++ src/test/java/animals/dog/Test1.java
@@ -1,22 +1,15 @@
-package animals;
+package animals.dog;
 
+import animals.Dog;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class DogTest {
-
-    Dog dog = new Dog();
+public class Test1 {
 
     @Test
     public void testDogVoice()
     {
+        Dog dog = new Dog();
         Assert.assertEquals(dog.getVoice(), "Гав!");
     }
-
-    @Test
-    public void testDogRun()
-    {
-        Assert.assertEquals(dog.getRun(), "Run");
-    }
-
 }
