+++ src/test/java/animals/dog/Test2.java
@@ -1,0 +1,16 @@
+package animals.dog;
+
+import animals.Dog;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+public class Test2 {
+
+    @Test
+    public void testDogRun()
+    {
+        Dog dog = new Dog();
+        Assert.assertEquals(dog.getRun(), "Run");
+    }
+
+}
